updated test for wallet with cert-manager option

+++ test/helpers/wallet.spec.ts
@@ -19,30 +19,55 @@ import { FileSystemWallet } from 'fabric-network';
 import * as fs from 'fs';
 import * as rimraf from 'rimraf';
 
+// tslint:disable-next-line: no-var-requires
+const CertificateManagerWallet = require('@blockchainlabs/ibm-certificate-manager-wallet');
 import * as walletHelper from '../../server/helpers/wallet';
 
-const cert = fs.readFileSync(`${__dirname}/../mocks/cert`, 'utf8');
-const key = fs.readFileSync(`${__dirname}/../mocks/key`, 'utf8');
+const cert = fs.readFileSync(`${__dirname}/../mocks/testuser1.pem`, 'utf8');
+const key = fs.readFileSync(`${__dirname}/../mocks/testuser1.key`, 'utf8');
+
+const FILESYSTEM_WALLET = 'FileSystemWallet';
+const CERTIFICATE_MANAGER_WALLET = 'CertificateManagerWallet';
 
 describe('helpers - wallet', () => {
+
     beforeAll(() => {
-        // delete test wallet before starting
-        rimraf.sync(config.get('fsWalletPath'));
+        // delete test wallet before starting for file system wallet
+        if (config.get('walletType') === FILESYSTEM_WALLET) {
+            rimraf.sync(config.get('fsWalletPath'));
+        }
     });
     afterAll(() => {
-        // delete test wallet after tests
-        rimraf.sync(config.get('fsWalletPath'));
+        // delete test wallet after tests for file system wallet
+        if (config.get('walletType') === FILESYSTEM_WALLET) {
+            rimraf.sync(config.get('fsWalletPath'));
+        }
     });
 
     let wallet;
 
     beforeEach(() => {
+        walletHelper.initWallet(config.get('walletType'));
         wallet = walletHelper.getWallet();
     });
 
+    describe('#initWallet', () => {
+        test('should fail for incorrect type of wallet', () => {
+            expect(() => {
+                walletHelper.initWallet('IncorrectWalletType');
+            }).toThrow(new Error('Incorrect walletType in config'));
+        });
+    });
+
     describe('#getWallet', () => {
-        test('should get wallet of type FileSystemWallet', () => {
-            expect(wallet).toBeInstanceOf(FileSystemWallet);
+        test('should get correct type of wallet', () => {
+            if (config.get('walletType') === FILESYSTEM_WALLET) {
+                expect(wallet).toBeInstanceOf(FileSystemWallet);
+            } else if (config.get('walletType') === CERTIFICATE_MANAGER_WALLET) {
+                expect(wallet).toBeInstanceOf(CertificateManagerWallet);
+            } else {
+                throw new Error ('Incorrect walletType in config');
+            }
         });
     });
